Consolidated Tables
I consolidated SSH and Telnet tables into one.
